Files and Images added as attachments
nstead of an array, just provide an url as filename. This may cause
warnings with message 'The variable or parameter item is empty.' or
'Unable to evaluate condition list_contains.'. This are generated
because the source property 'files' is empty.

Also changed the regex to case insensitive, something the files are
referenced by <a href="…"</a> or <A HREF="…."/>.

Also added similar code for images that are attached. Regex is somewhat
different because of the '>' character.